requirements.txt
numpy==2.1.3
matplotlib==3.9.2
pandas==2.2.3
python-dotenv==1.0.1
scikit-learn==1.5.2
scipy==1.14.1
statsmodels==0.14.4